refactor: 멤버십 플랜 모달 이름 변경
